feat: admin api -> lottery api
