+++ client/src/router.tsx
@@ -1,5 +1,4 @@
 import { createBrowserRouter } from "react-router-dom";
-import { AdminAPI } from "./apis/adminAPI";
 import { LotteryAPI } from "./apis/lotteryAPI";
 import Layout from "./components/Layout";
 import CasperCustom from "./pages/CasperCustom";
@@ -27,7 +26,7 @@ export const router = createBrowserRouter([
                     {
                         index: true,
                         element: <Lottery />,
-                        loader: AdminAPI.getLottery,
+                        // loader: LotteryAPI.getLottery,
                     },
                     {
                         path: "custom",
