--- client/src/types/adminApi.ts
@@ -1,6 +1,0 @@
-export type GetLotteryResponse = {
-    lottery_event_id: number;
-    start_date: string;
-    end_date: string;
-    winner_count: number;
-}[];
