+++ client/src/types/lotteryApi.ts
@@ -21,3 +21,10 @@ export interface PostCasperResponse extends CasperInformationType {
 export interface GetApplyCountResponse {
     appliedCount: number;
 }
+
+export type GetLotteryResponse = {
+    lottery_event_id: number;
+    start_date: string;
+    end_date: string;
+    winner_count: number;
+}[];
